refactor(bao): reorganize client and config code
- Move Client type and NewClient function from config.go to client.go
- Keep config.go focused on Config and RetryConfig types only
- Update error messages from 'vault:' to 'bao:' prefix
- Clarify SDK quirk comment about nil response handling
- Update documentation to reference 'bao' instead of 'vault'
- All tests pass (30.6s with race detection)
